Create README.md
I'm interested in being a friend of the felines. 
I'm currently learning spirituality and teying to start tapping into my psychic abilities.
I'm looking to see into the future, trusting my intuition, going with the flow,  listening to the universe and living my best life. If I can collaborate with psychics, intuitive and mind readers for some inside information to help me along the way, that'd be much much appreciated than I can grow and be a better human being.